Refactor: Reduce some repeated code
